+++ docs/examples/ButtonExample.vue
@@ -1,10 +1,16 @@
 <template>
-    <div class="width-capped-content">
+    <div class="width-capped-content vertical-layout">
         <!-- #region example -->
         <ipl-button
             label="Button"
             :disabled="disabled"
             :small="small"
+        />
+        <ipl-button
+            label="Button"
+            :disabled="disabled"
+            :small="small"
+            class="no-text-transform"
         />
         <!-- #endregion example -->
     </div>